+++ backend/pkg/scraper/scraper.go
@@ -1,1 +1,3 @@
 package scraper
+
+// Scraper defines the interface for a scraper
