+++ backend/pkg/store/store.go
@@ -1,0 +1,4 @@
+package store
+
+// SaveArticles saves articles to a JSON file
+func SaveArticles() {}
