+++ frontend/components/NavBar.tsx
@@ -11,16 +11,8 @@ const navLinks = [
     path: "/",
   },
   {
-    name: "Blog",
-    path: "/blog/",
-  },
-  {
     name: "About",
     path: "/about/",
-  },
-  {
-    name: "Contact",
-    path: "/contact",
   },
 ];
